fix: change main file to newer
